ZEPPELIN-4176. Remove old spark interpreter
### What is this PR for?

This PR is just to remove the old spark interpreter. The old spark interpreter has several issues, and we introduce new spark interpreter implementation in 0.8. This ticket is to remove it in 0.9. Here's the issues of old spark interpreter.

- Didn't use native scala shell api.
- Dependency management is not applied for yarn cluster mode.
- It can not support scala 2.12 due to point 1

### What type of PR is it?
[ Improvement ]

### Todos
* [ ] - Task

### What is the Jira issue?
* https://jira.apache.org/jira/browse/ZEPPELIN-4176

### How should this be tested?
* CI pass

### Screenshots (if appropriate)

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: Jeff Zhang <zjffdu@apache.org>

Closes #3375 from zjffdu/ZEPPELIN-4176 and squashes the following commits:

4efa61fa7 [Jeff Zhang] address comment
0578f4aaa [Jeff Zhang] ZEPPELIN-4176. Remove old spark interpreter
